+++ src/main/java/com/lanagj/adviseme/recommender/nlp/weight/co_occurrence_matrix/WordOccurrenceMatrix.java
@@ -10,5 +10,5 @@ public interface WordOccurrenceMatrix<T> {
      * @param documentContextMatrix key - document ID, value - list of words from document
      * @return key - word, value - calculated word frequency
      */
-    Map<String, List<T>> get(Map<Long, List<String>> documentContextMatrix);
+    Map<String, List<T>> get(Map<Integer, List<String>> documentContextMatrix);
 }
